# Optimization for MS search