Pieguy pointed out that we can eliminate a large chunk of the search
space due to how MSCC chooses blob directions.

Benchmarks of MS search only (TW search commented out):

Before:
    searched 100001 seeds in 1655.809000 ms
    average 16.6 us/seed

After:
    searched 100001 seeds in 744.238000 ms
    average 7.4 us/seed